Add pots to send midi CC messages and alsways MGB mode
I added 3 potentiometers to control the waveshape for the
three (melodic) channel and a fourth potentimeter controls the
wave offset for the third channel.
I have also changed some code to always be in MGB mode, because that is
all I use and I don't have LEDs or buttons on my arduino (nano).
Code is very hacky, sorry.
All potentiometers are set up as voltage dividers between 5V and 0V.